Modified PoolCatalog.cc to take the catalog contact string from the site-local config file if that exists. If it doesn't, it uses file:PoolFileCatalog.xml. This will allow us to begin deploying and using the site-local config file, but without yet requiring it. At some point we can make the change to require it.
Note that I also put in a change to require that the envvar CMS_PATH be
set. This _should_ be the case for every environment, but we will see. I'd
rather know now if this isn't the case for some strange reason.

Still to do: put in the code to read the real site-local config file (an
xml file format).